+++ internal/adapters/api/handler/role.go
@@ -9,16 +9,14 @@ import (
 )
 
 type roleHandler struct {
-	roleService       ports.RoleService
-	permissionService ports.PermissionService
-	logger            ports.Logger
+	roleService ports.RoleService
+	logger      ports.Logger
 }
 
-func NewRoleHandler(roleService ports.RoleService, permissionService ports.PermissionService, logger ports.Logger) ports.RoleHandler {
+func NewRoleHandler(roleService ports.RoleService, logger ports.Logger) ports.RoleHandler {
 	return &roleHandler{
-		roleService:       roleService,
-		permissionService: permissionService,
-		logger:            logger,
+		roleService: roleService,
+		logger:      logger,
 	}
 }
 
@@ -38,9 +36,14 @@ func (h *roleHandler) CreateRole(c *gin.Context) {
 
 func (h *roleHandler) GetRoleByID(c *gin.Context) {
 	id := c.Param("id")
+	if !domain.IsUUID(id) {
+		response.JSON(c, "invalid_request", http.StatusBadRequest, nil, nil)
+		return
+	}
+
 	role, err := h.roleService.GetRoleByID(id)
 	if err != nil {
-		response.JSON(c, "failed to find role", http.StatusNotFound, nil, []string{err.Error()})
+		response.JSON(c, "failed to find role", http.StatusInternalServerError, nil, []string{err.Error()})
 		return
 	}
 	response.JSON(c, "success finding role", http.StatusOK, role, nil)
@@ -49,7 +52,7 @@ func (h *roleHandler) GetRoleByID(c *gin.Context) {
 func (h *roleHandler) GetAllRoles(c *gin.Context) {
 	roles, err := h.roleService.GetAllRoles()
 	if err != nil {
-		response.JSON(c, "invalid_input", http.StatusNotFound, nil, []string{err.Error()})
+		response.JSON(c, "failed to find roles", http.StatusInternalServerError, nil, []string{err.Error()})
 		return
 	}
 	response.JSON(c, "success retrieving roles", http.StatusOK, roles, nil)
@@ -57,6 +60,11 @@ func (h *roleHandler) GetAllRoles(c *gin.Context) {
 
 func (h *roleHandler) DeleteRole(c *gin.Context) {
 	id := c.Param("id")
+	if !domain.IsUUID(id) {
+		response.JSON(c, "invalid_request", http.StatusBadRequest, nil, nil)
+		return
+	}
+
 	if err := h.roleService.DeleteRole(id); err != nil {
 		response.JSON(c, "failed to delete role", http.StatusInternalServerError, nil, []string{err.Error()})
 		return
